+++ src/requests/request-manager.ts
@@ -1,8 +1,9 @@
 // Copyright (C) Microsoft Corporation.  All rights reserved.
 
 import { DebugProtocol } from '@vscode/debugprotocol';
-import { DebuggerRequestArguments, DebuggerRequestEnvelope, DebuggeeResponseEnvelope } from './debugger-request-schema';
+import { DebuggerRequestArguments} from './debugger-request-schema';
 import { IDebuggeeMessageSender } from '../debuggee-message-sender';
+import { DebuggeeResponseEnvelope, DebuggerRequestEnvelope } from '../protocol-events';
 
 interface PendingDebuggerRequest {
     resolve: (value: DebuggeeResponseEnvelope) => void;
